// apps/admin/src/types/grobal-type.ts
export interface ICardVo {
  email: string
  fileName: string
  fileURL: string
  login: boolean
  msg: string
  name: string
  phone: string
  rank: string
  team: string
  telephone: string
  theme: string
  code?: string
}

export interface IWorkVo {
  contents: string
  time: number
  title: string
}